prevent trashing when building with rustc in Jetbrains IDEs.
For the IDE to discover what's going on, it will use hidden environment variables
to perform the build with, but that also affects certain crates which then report
they changed.

If this is intermixed with actual nightly builds, unnecessary rebuilds happen all the time.
To prevent this, just use `nightly` by default without actually using nightly features.