Update Ownership class
Rename the `Ownership` class to `SelectedOwnership`, as it is only used to determine the selected state of the files to add to a commit.

Renamed the methods as well to convey their actual purpose.

The `SelectedOwnership` class will be updated in a way that the selected is persisted across file/hunk ownership updates