Assure that `Project::path` can be compared reasonably with other paths (#5096)
When testing paths for prefix-matches it's important they are all normalized in
the same fashion.

`canonicalize()` is very particular about canonicalizing Windows paths, which
makes it easy for these paths to not be compatible to other absolute-looking
paths.

The difficulty here is to get the right trade-off between performance and
safety, e.g. we wouldn't want these canonicalized Windows paths to be used
anywhere as they are very uncommon (and don't even work everywhere).