fix: do not cutoff traversals by date for branch details (#1650).
We already found the merge-base and know that this will naturally
be the end of the traversal. Thus there is no need to additionally
set a cut-off date, which is the documented default.

With such a cutoff, it's very possible that certain operations on the branches
change the date of the base to lie in the future, which is when no traversal
would be able to happen anymore.