Improvew is integrated check
This commit does two things to improve the is_integrated check.

Firstly, it uses passes true to the log functions. This means that all branches of merge commits will be explored in the check for upstream integration.

It also now uses the change ID to see if a commit has been integrated.