Clean Kiri's commits
We now do error handling at a global level, so those the try catches are not helpful anymore.

The `!$user` check is presumably a reminant from when we had the admin check performed in those funcitons. Instead we have the admin check before the user can even see this page