produce input stacks from workspaces for use in lock dependency computation.
CLI Integration
---------------

It's also possible to run the dependency engine through the CLI which should help debugging.

Refactors
---------

Prevent `gitbutler-branch-actions` tests to run twice as
they were running as individual tests.

Hold back on refactoring the test layout though, as this is coming
with the `but-*` crates naturally.

Also port all tests.