Added two feature flags to guide users away from incomplete features
In an effort to smooth out the initial user experience, I've hidden some pages that the user would otherwise have to page through in order to make use of the currently implemented features.

Namly, this hides the organizations page (which is currently hard to understand & mostly unused) as well as the project page. In the future the project page will be similar to the GH repo page.

This commit also corrects the link to the `/profile` page. Previously it was linking to the nonexistant `/settings/profile` page, which due to it's unforutnate bout of nonexistance was getting interpreted as a project instead.

The experimental settings are toggle-able on the now accessable `/profile` page