match on a patch url and use a different image for that
this server hook code now will see if it's a patch we're linking to and provide the new specific OG image for it.

If it's a branch review link, we'll attempt to fetch() the metadata and properly inject it so the unfurl is better.

We can't yet do this for the patch, because we don't have an unauthenticated API endpoint for that yet.