Push stack to review takes "topBranch" param
This commit allows the caller to specify what the top-most branch to be pushed should be.

The top-most branch should be one of the BranchHead names from the stack getting pushed.

All the branches including and beneath the top-most branch get pushed and uploaded to the GitButler server. This is a sensible behaviour as otherwise managing the lastPushedBase would end up being all over the place, and the stacked branches would look all off-whack.