Have publish BR buttons on each branch header & prevent getting stuck in "not found" state
This took quite a quile to figgure out the best coures of action.

The `SeriesHeader` component is VERY messy, and should be a candidate for refactoring in the near future. I wanted to shape the code in such a way that we can start having the combined PR&BR card so the comprimise that I came up with was the `BranchReview` component that has slots where the existing `PullRequestCard` slots in, with all of it's many many properties.

## What PR/BR buttons to show?

In the design, we have this multi-state button that can go down to just one state. To achieve this I am dynamically pushing the valid states into an array, and then rendering both in a dropdown button, or the one remaining in a regular button, where we are still able to refer to the "selectedAction" which gets set to the first available action by default.

## Dealing with the But Request polling time

One "issue" with the but request upload mechanism is that the processing happens asynchronusly. This results in a scenario where the frontend knows about the reviewId (because the frontend gets to deside the reviewId), but the backend doesn't know about it.

To deal with this, I've added in some basic rapid (IE faster than the usual 5 minute polling) polling that will trigger a re-fetch with exponential back off.

The code to implement this looks a bit strange. Trying to manage a longer lived async processes is a bit of a pain point in svelte. When I've found the branch, or if svelte looses interest, I want to abort the polling. To achieve this I pass in an object to the `pollWhileNotFound` function, which because of JS passing things by "value reference" essentially lets me share some memory between the containing effect and the function, and by setting a certain value to false in the effect, I can signal that the polling should be stopped.