Use entity adapter with uncommitted changes API
Using an entity adapter gives us the ability to`selectById` for free, and ensures the data is sorted.

Separately we also ignore the ignored files that is part of the worktree changes response. In conversation with Caleb we came to the conclusion it would be best to include them in the same list as normal changes, but treat them differently like we already do for e.g. binary files.