Remove `Merge` Step in `but-rebase` to remove `gitbutler-repo` dependency.
There is no use of it yet and it doesn't look like it's needed
since `Pick` can re-merge merge-commits.

With the dependency gone, the new `but-rebase::cherry_pick` can be used
in `gitbutler-repo` to assure all tests still work.

From there certain crucial tests can be ported to `but-rebase` as
cherry-pick tests.