Allow cherry-picks to handle merge commits (#7396)
Now that the 'old' code uses the new cherry-pick just for good measure,
it needs to be functionally equivalent. This also means that it has to
be able to cherry-pick merge commits, something which should just work.

The re-merge we can only do if it's our workspace commits, but even then
a cherry-pick should be enough. One step at a time though.

NOTE: this doesn't trivially work.