Return more infomation after upstream integration
We want to enable the frontend to communicate with other services depending on what happens in the integration step.

The first thing that we are interested in is making butler requests as closed whenever they have been integrated with target (origin/master) branch. As such, this starts a new struc and provides the archived_branches property, which is specifically the newly archived branches.