Fix startup loop
In the base page, we have something that reacts to changes in the list of all projects, and tries to redirect to one of those projects.

Unfortunately, if we update hte list of all projects inside of the load function, the base page gets ahead of us, and actually triggers a second navigation call, which in turn updates the list of projects, ad infinitum. 

We _do_ want to react to the list of projects in the base page, so we should instead consider not updating the list inside of the load function (and in generally, having the load function cause side effects is not a nice thing).

---

 I've firstly made `setActiveProject` a method on the `ProjectsService` and tied the reload to it, so that is all encapsulated in one call.

To actually fix the issue, I've moved the call to `setActiveProject` outside of the load function, and put it in an `afterNavigate` call, which has a little extra logic, to make sure it's only run whenever the projectId in the path changes.