Use the value of `SHELL` for the shell, if available. (#7467)
The `SHELL` environment variable is set on MacOS to represent the login shell.
It is probably more likely to yield an environment that can run the signing program.

Note that doing so usually is dangerous as well if scripts were to be used,
but that's not the case here - we know a program is provided.