Assure that the commit-lines compare to the same 'conversion'
When comparing hunks we must be sure these refer these have been created
from the same content.

Currently, `unified_diff()` produces the content by applying binary-to-text filters,
which is good for presentation of binary files, but bad if hunks actually get selected there.

For now, rely on hunks having to match perfectly to be applicable, but assure that the content
used for file generation is the same (so the hunks that match actually refer to the right thing).