Increase time given for login attempt
This PR dramatically increases the amount of time users have when trying to login.

Previously there was only 2 minutes which is not a lot of time to complete 2fa or other steps like that.

This also introduces the ability to abort the login attempt because if the user determines that it has failed for whatever reason, it would be daft for them to wait 20 minutes to go again.