Handle branch removal in v3
This adds forked logic to handle branch unapplying in gitbutler v3.

The main change here is to move away from tree based manipulations.

We need to handle the uncommited  changes (cwdt) as just one tree. We want to somehow remove the contents of one branch from this tree. As it turns out, this is pretty much analogus to what the `git revert` command does.

TBH, I've not read how `git revert` is implemented, but I've chosen to achieve the behaviour using a 3 way merge.

There are three relevant trees to think about when performing this operation. The `cwdt`,  the head of the branch to remove (`rh`), and the workspace base (`wb`).

If we diff `cwdt` against `rh`, we get a diff that introduces all the changes that the other branches introduce. If we  diff `wb` against `rh`, we get a diff that removes all of the contents of the branch to remove. Combining those together in a three way merge results in `cwdt` without the branch we want to remove