Use separate selection lists for uncommitted, commits and branches
We don't want the uncommitted changes selection to clear when you select a file
from a commit, and further, we want to preserve selections per commit.

Not addressed in this change is clearing things that are no longer relevant, but
it feels easy enough to add some periodic cleanup based on recent usage.
Temp commit