Merge the local and remotes commits with the stack info
The newly named stack details endpoint gets the information about the given stack. Since this information heavily depends on information derived from the commits, we just return the commit information as well.

This spares us having to make a second API call to the now-dead local and remote commits endpoint.

This also spares us the need to track what endpoints to invalidate