Richard fix: text wrapping
The UI would freeze because it would basically fall into a endless loop of trying to wrap the words around when there would be a word larger than the wrap value.

This is no longer the case.

Instead, the word will be broken only if it alone surpases the wrap value.