multi-lanes-wip (#8445)
* feat(ui): Adjust status badge padding for better alignment

Reduces the top and bottom padding of the status badge in the
IntegrationSeriesRow component to create a more visually balanced
appearance and improve the overall layout.

* feat(ui): Enhance file status badge with stroke width

Increase the stroke width of the file status badge to improve
its visibility and overall design. This change ensures the
badge stands out more clearly against the background,
providing a better user experience.

* feat(ScrollableContainer): add children wrapper and scrollbar

In order to fix broken “on-hover” mode

* update lanes view

- new header styles
- fix vertical and horizontal scrollbars
- update lane styles a little

* improve vertical mode

* resolve merge conflicts

* feat(MultiStackPagination): add optional activeIndex prop

* Drop some of the “Tabs” code to get through linter errors