Add two global variables for console debugging
Having access to `uiState` and `idSelection` from the console makes
debugging certain things a bit more convenient.

I tried to think of risks related to doing this, but could not come up
with any. We should perhaps just avoid using these for testing
purposes, it's better to test public interfaces.