adjust tests expectations to match previous implementation.
This is a loss in functionality for submodule checkouts, but a win for
removing hunks as now it will not undo the executable bit changes.
These are then left for a second discard, as such differences would show
up as worktree changes.

Conflicts are once again ignored, but a note was added to explain why
that's preferable for now.