assure base-commits are the same between branch-listing details and branch-details
Otherwise the branch-listing shows there are three commits, but the branch view shows
there are 900.

Turns out that using the integration branch as merge-base for branches is desired for now,
and it matches what the branch listing does as well.
However, we noticed that it's difficult to see what a branch is compared against sometimes.