Speculatively fix eslint config by upgrading packages
There is something wonky about the parsing of typescript/svelte files, 
the main symptom being "go to defintion" not working. This has other 
consequences, such as renaming variables producing garbage in files that
references it.

It is unclear why upgrading these pacakges should fix it, but locally it
seems to do the job.