Test combine results
This fixes at two errors:

When no results were passed, an error was caused

Not correctly passing the error through

Now when you pass zero results as an input, it returns an uninitialized result as the outcome.

I also noticed when testing, that the output error property was getting set to the entire result that contained the error, which is probably where a handful of [object Object]s might have come from.